Dialog section Code Modular, Break Dialog Section #2114
