--- src/components/Auth/ForgotPassword/ForgotPassword.css
@@ -1,20 +1,0 @@
-.forgotPwdForm{
-	margin: 0 auto;
-	padding: 2rem;
-}
-
-.forgotPwdForm h3{
-	font-family: 'Open Sans', sans-serif;
-	margin: 5px 0;
-	font-weight: 500;
-}
-
-.forgot:hover {
-  text-decoration: underline;
-}
-
-@media screen and (max-width: 768px) {
-	.forgotPwdForm h3{
-		font-size: 16px;
-	}
-}
